Set song title from media player broadcast.

+++ skipshuffle/src/main/java/com/dontbelievethebyte/skipshuffle/RandomPlaylist.java
@@ -68,7 +68,7 @@ public class RandomPlaylist implements PlaylistInterface {
 
     @Override
     public Track getCurrent() {
-        return dbHandler.getTrack(_tracksIds.get((int) playlistPosition));
+        return dbHandler.getTrack(_tracksIds.get(playlistPosition));
     }
 
     @Override
